Implementing get methods for property to implement a check if property exists

+++ job-card/src/main/java/com/dzt/job_card/Models/Property.java
@@ -156,7 +156,7 @@ public class Property {
         return acreage;
     }
 
-    public void setAcreage(int acreage) {
+    public void setAcreage(double acreage) {
         this.acreage = acreage;
     }
